feat: add port from .env
