+++ .env
@@ -1,4 +1,4 @@
-PORT=4000
+PORT1=4000
 
 CRYPT_SALT=10
 JWT_SECRET_KEY=secret123123
