+++ src/main.ts
@@ -1,9 +1,11 @@
 import { NestFactory } from '@nestjs/core';
 import { AppModule } from './app.module';
 import { DocumentBuilder, SwaggerModule } from '@nestjs/swagger';
-
+import 'dotenv/config';
 async function bootstrap() {
   const app = await NestFactory.create(AppModule);
+
+  const PORT = process.env.PORT1 || 4000;
 
   const docConfig = new DocumentBuilder()
     .setTitle('Home Library Documentation')
@@ -14,6 +16,6 @@ async function bootstrap() {
   const document = SwaggerModule.createDocument(app, docConfig);
   SwaggerModule.setup('doc', app, document);
 
-  await app.listen(4000);
+  await app.listen(PORT);
 }
 bootstrap();
